Crate cairus

The main crate for Cairus.

Overview

Cairus is a 2D graphics library based on the Cairo vector graphics library. Cairus is designed to utilize and preserve the Cairo drawing model while providing the benefits of a native Rust implementation.

Modules

context

operators

surfaces