Struct cairus::context::Context

pub struct Context<'a> {
    pub rgba: Rgba,
    // some fields omitted
}

Struct defined for context

Fields
rgba: Rgba
Methods

impl<'a> Context<'a>

Implementation of methods for context

fn create(target: &'a mut ImageSurface) -> Context

fn set_source_rgba(&mut self, red: f32, green: f32, blue: f32, alpha: f32)

Sets Rgba values of source to used defined values This function changes the Rgba values of the source

fn paint(&mut self)

Paints this context's Rgba on the destination surface with the over operator.

This is a completely naive, and frankly useless implementation. It is a place holder for the real paint function to later be implemented. It operates on every 'pixel' of the destination surface.