Function cairus::operators::fetch_operator

pub fn fetch_operator(op: &Operator) -> fn(_: &Rgba, _: &mut Rgba)

Returns an image compositing function that corresponds to an Operator enum.

This function maps an enum to its function, allowing for dynamic determination of the operator function. This is likely a good way for a context to fetch the correct function just by having an Operator enum, instead of it having to use a match statement to find the correct operator, or having this fetch function implemented in the context module (away from the rest of the operator definitions).

Arguments

op - Reference to an enum Operator that matches the desired operation.

Usage

let op_enum = Operator::Over;

// Fetch and use the operator let compose = fetch_operator(&op_enum); compose(&source, &mut destination1);