Function cairus::operators::operator_in

pub fn operator_in(source: &Rgba, destination: &mut Rgba)

This is Cairus' in operator. The destination object is removed and the source object is only drawn where the destination was. Note: The transparency of the first object is still taken in to account. The effect of the IN operator depends on the interpretation of the source. This operator is unbounded. This function currently assumes post-multiplied alpha values, the alpha value must be factored out