Function cairus::operators::operator_over

pub fn operator_over(source: &Rgba, destination: &mut Rgba)

Operator Formulas

The following functions are implementations of the Porter Duff operator formulas. (See below for the Porter Duff paper in the references section, or the Cairo operator documentation page). Composites source over destination.

Arguments

source - The source Rgba to be applied to the destination Rgba.

destination - The destination Rgba that holds the resulting composition.

Over is Cairus's default operator. If the source is semi-transparent, the over operation will blend the source and the destination. If the source is opaque, it will cover the destination without blending. Assumes pre-multiplied alpha.