Module cairus::surfaces

Overview

Cairo surfaces are basically raster (bitmap) containers. They 'receive' operations performed on them by contexts. They are the 'canvas' of Cairus.

Structs

ImageSurface
A surface needs to hold pixels (Rgba's) and its width and height. The width and height will be used in rendering to images and calculating clipping, and the pixels will be the things that actually are operated on by stroke or paint operations. See the test_image_surface_with_operator test case below for an example of what that might look like.

Enums

Format
Format enum descriptors for the surface object These are specifically the format types copied from the C implementation, analogous to cairo_format_t

Type
Analogous to cairo_surface_type_t, indicates target drawing type