Struct cairus::surfaces::ImageSurface

pub struct ImageSurface {
    pub width: usize,
    pub height: usize,
    // some fields omitted
}

A surface needs to hold pixels (Rgba's) and its width and height. The width and height will be used in rendering to images and calculating clipping, and the pixels will be the things that actually are operated on by stroke or paint operations. See the test_image_surface_with_operator test case below for an example of what that might look like.

Fields
width: usize height: usize
Methods

impl ImageSurface

ImageSurface provides iter(), into_iter(), and iter_mut() so that when a Cairus context calls paint, it can simply iterate through the pixels in the image surface and use a image compositing operator to operate on them. See operators.rs for those operations. A valid ImageSurface needs tangible height and width dimensions to be successfully created.

fn create(width: usize, height: usize) -> ImageSurface

fn iter(&self) -> Iter<Rgba>

fn iter_mut(&mut self) -> IterMut<Rgba>

fn to_file(&self, path: &Path)

The external Rust Image Crate is a library under development to read, manipulate and write images. At the moment "image" supports reading and writing JPG and PNG images. The below functions, to_file(), to_png, and to_jpg use this external library to write output image files, provided a valid Cairus ImageSurface.

fn get(&self, x: usize, y: usize) -> Option<&Rgba>

fn get_mut(&mut self, x: usize, y: usize) -> Option<&mut Rgba>

fn get_with_index(&self, idx: usize) -> Option<&Rgba>

fn get_mut_with_index(&mut self, idx: usize) -> Option<&mut Rgba>